PRSD-835: Moves form-related view models into formModels
