+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/pages/SelectLocalAuthorityPage.kt
@@ -3,7 +3,7 @@ package uk.gov.communities.prsdb.webapp.forms.pages
 import org.springframework.ui.Model
 import org.springframework.validation.Validator
 import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.SelectLocalAuthorityFormModel
-import uk.gov.communities.prsdb.webapp.models.viewModels.SelectViewModel
+import uk.gov.communities.prsdb.webapp.models.viewModels.formModels.SelectViewModel
 import uk.gov.communities.prsdb.webapp.services.LocalAuthorityService
 
 class SelectLocalAuthorityPage(
