+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/formModels/CheckboxViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.formModels
 
 data class CheckboxViewModel<T>(
     val value: T,
