+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/formModels/RadiosViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.formModels
 
 abstract class RadiosViewModel(
     open val labelMsgKey: String? = null,
