+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/formModels/SelectViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.formModels
 
 data class SelectViewModel<T>(
     val value: T,
